pyproject.toml
[project]
name = "teste-redspark"
version = "0.1.0"
description = "Relatorio Aprovados RPA"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv-python>=0.0.1",
    "faker>=37.0.0",
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
]
